Alsa output driver
Based on a bit of code of @Skaman and bit of mine, I've added support
for mixer and audio Alsa support. I'm no master in C (or any
programming language for that matter), but I figured it would be a
useful job for this sunday afternoon. In the course of creating/testing
it I've come across a few bugs, which I think are not related to this
code. I'll address them later, perhaps with a patch if it's within my
abilities.