Fix data allignment issue
Different processor targets may have different word alignment restrictions.
Typically, a two or four byte integer should start on a four-byte address in
order to be fetched correctly over a bus. An access of a misaligned memory
element may cause a bus error.

Therefore, we need to adjust the contents of a message by copying it into a new
location, before interpreting it.