mdns: avahi: set AVAHI_CLIENT_NO_FAIL in avahi_client_new()
For whatever reasons `avahi_client_new` passes the value `0` as its
second parameter. However, the second parameter is of type
[`AvahiClientFlags`](http://avahi.org/doxygen/html/client_8h_source.html#l00048)
with the possible values:

``` c
typedef enum {
    AVAHI_CLIENT_IGNORE_USER_CONFIG = 1,
    AVAHI_CLIENT_NO_FAIL = 2
} AvahiClientFlags;
```

Setting `AVAHI_CLIENT_NO_FAIL` has the benefit that the shairport-sync
daemon does not depend on the order of the init process whether the
avahi daemon is already up.